Add bugfixing logic
I added automaticallyVerticalDirection var and some logic for it. If this var equals true, tooltip will be shown up (if the calling place is located below the center by axis Y) and tooltip will be shown down (if the calling place is located above the center by axis Y). 
This logic let to avoid some UI bugs and the similar logic is used on other widgets (dropdowns, popups, tooltips and etc).